requirements.txt
opensearch-py>=2.8.0
opensearch-py-ml>=1.1.0
PyYAML>=6.0
boto3>=1.28.0